backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
httpx>=0.24.1
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
